Update README.md
This repository contains a machine learning model that predicts the power consumption of a household based on various parameters such as temperature, humidity, and time of day. The model is implemented using Python and the scikit-learn library.

Dataset
The dataset used for this project is the Household Power Consumption dataset from the UCI Machine Learning Repository. The dataset contains data for one household over a period of 4 years.

Model
The model uses a Random Forest Regressor for prediction. The parameters used for prediction are temperature, humidity, and the hour of the day. The model is trained on 70% of the data and tested on the remaining 30%.

Usage
The code is written in Python 3.6 and uses the following libraries:

numpy
pandas
matplotlib
seaborn
scikit-learn
To run the code, simply clone the repository and run the household_power_consumption.ipynb Jupyter Notebook file. The code will automatically download the dataset, preprocess it, and train the model.

Results
The model has an accuracy of 92.45% on the test data. The results can be visualized by running the code in the Jupyter Notebook file.

Conclusion
The model can be used to predict the power consumption of a household based on temperature, humidity, and time of day. The model can be further improved by incorporating other factors such as weather conditions, holidays, and other household events.